Remove all related to feature flag
