+++ web/src/index.jsx
@@ -11,8 +11,7 @@ import { store, sagaMiddleware } from './store';
 import { cordovaApp } from './cordova';
 import sagas from './sagas';
 
-import Main from './layout/main_old';
-import MainForHistoricalView from './layout/main';
+import Main from './layout/main';
 import GlobalStyle from './globalstyle';
 
 // init styling
@@ -36,11 +35,6 @@ console.log(
 // Plug in the sagas
 sagaMiddleware.run(sagas);
 
-// Switches Main component if historical view feature google is used
-const urlSearchParams = new URLSearchParams(window.location.search);
-const isHistoryFeatureEnabled = Object.fromEntries(urlSearchParams.entries())?.feature === 'history';
-const MainComponent = isHistoryFeatureEnabled ? MainForHistoricalView : Main;
-
 // Render DOM
 ReactDOM.render(
   <React.Suspense fallback={<div />}>
@@ -49,7 +43,7 @@ ReactDOM.render(
       {/* the route history outside of React components anymore */}
       <Router history={history}>
         <GlobalStyle />
-        <MainComponent />
+        <Main />
       </Router>
     </Provider>
   </React.Suspense>,
