+++ web/src/layout/map.jsx
@@ -12,7 +12,6 @@ import { debounce } from '../helpers/debounce';
 import { useInterpolatedSolarData, useInterpolatedWindData } from '../hooks/layers';
 import { useCo2ColorScale, useTheme } from '../hooks/theme';
 import { useZonesWithColors } from '../hooks/map';
-import { useFeatureToggle } from '../hooks/router';
 import { dispatchApplication } from '../store';
 
 import ZoneMap from '../components/zonemap';
@@ -45,7 +44,6 @@ export default () => {
   const location = useLocation();
   const history = useHistory();
   const co2ColorScale = useCo2ColorScale();
-  const isHistoryFeatureEnabled = useFeatureToggle('history');
 
   // TODO: Replace with useParams().zoneId once this component gets
   // put in the right render context and has this param available.
@@ -193,7 +191,6 @@ export default () => {
       <ZoneMap
         co2ColorScale={co2ColorScale}
         hoveringEnabled={hoveringEnabled}
-        isHistoryFeatureEnabled={isHistoryFeatureEnabled}
         onMapLoaded={handleMapLoaded}
         onMapError={handleMapError}
         onMouseMove={handleMouseMove}
