+++ web/src/layout/timeController.jsx
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import TimeSlider from '../components/timeslider_new';
 
 import { useCurrentDatetimes } from '../hooks/redux';
-import { useFeatureToggle } from '../hooks/router';
 import { dispatchApplication } from '../store';
 import styled from 'styled-components';
 
@@ -62,11 +61,6 @@ const StyledTimeSlider = styled(TimeSlider)`
 
 const TimeController = ({ selectedZoneTimeIndex, selectedTimeAggregate }) => {
   const datetimes = useCurrentDatetimes();
-  const isHistoryFeatureEnabled = useFeatureToggle('history');
-
-  if (!isHistoryFeatureEnabled) {
-    return null;
-  }
 
   const startTime = datetimes.at(0);
   const endTime = datetimes.at(-1);
