+++ web/src/reducers/index.js
@@ -4,7 +4,6 @@ import { getKey } from '../helpers/storage';
 import { isLocalhost, isProduction } from '../helpers/environment';
 
 import dataReducer from './dataReducer';
-import dataReducerForHistoryFeature from './dataReducerForHistoryFeature';
 import { TIME } from '../helpers/constants';
 
 const getStorageBool = (key, defaultValue) => {
@@ -79,11 +78,7 @@ const applicationReducer = (state = initialApplicationState, action) => {
   }
 };
 
-const urlSearchParams = new URLSearchParams(window.location.search);
-const isHistoryFeatureEnabled = Object.fromEntries(urlSearchParams.entries())?.feature === 'history';
-const dataReducerInUse = isHistoryFeatureEnabled ? dataReducerForHistoryFeature : dataReducer;
-
 export default combineReducers({
   application: applicationReducer,
-  data: dataReducerInUse,
+  data: dataReducer,
 });
