+++ web/src/sagas/index.js
@@ -6,11 +6,7 @@ import { getGfsTargetTimeBefore, getGfsTargetTimeAfter, fetchGfsForecast } from 
 
 function* fetchZoneHistory(action) {
   const { zoneId, features, selectedTimeAggregate } = action.payload;
-  let endpoint = `/v4/history?countryCode=${zoneId}`;
-
-  if (features.includes('history')) {
-    endpoint = `/v5/history/${selectedTimeAggregate}?countryCode=${zoneId}`;
-  }
+  let endpoint = `/v5/history/${selectedTimeAggregate}?countryCode=${zoneId}`;
 
   if (features.length > 0) {
     endpoint += `${features.map((f) => `&${f}=true`)}`;
@@ -27,11 +23,7 @@ function* fetchZoneHistory(action) {
 
 function* fetchGridData(action) {
   const { features, selectedTimeAggregate } = action.payload;
-  let endpoint = '/v4/state';
-
-  if (features.includes('history')) {
-    endpoint = `/v5/state/${selectedTimeAggregate}`;
-  }
+  let endpoint = `/v5/state/${selectedTimeAggregate}`;
 
   if (features.length > 0) {
     endpoint += `?featureflag=true${features.map((f) => `&${f}=true`)}`;
